+++ v3/main.py
@@ -12,7 +12,7 @@ from db.db_models import Missions, GroupMissions
 from peft import PeftModel
 from datetime import datetime
 from tool.get_week_index import GetWeekIndex
-from core.core_tool import get_top_posts, get_group_info, largest_mission_id
+from core.data_tools import get_top_posts, get_group_info, largest_mission_id
 from core.llm_tools import (
     create_gen_llm_chain,
     create_eval_llm_chain,
